Arranged the files a little bit better

+++ src/java/Controllers/ProcessPollsController.java
@@ -1,4 +1,5 @@
 package Controllers;
+
 import Services.PollService;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
@@ -18,20 +19,34 @@ public class ProcessPollsController extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
 
+        String header = request.getHeader("requestType");
         String id = request.getParameter("id");
+
         try {
-            
-            pollService = new PollService();
-            
-            boolean didDelete = pollService.deletePoll(id);
-            
             JSONObject obj = new JSONObject();
-            obj.put("deleted",didDelete);
+            pollService = new PollService();
+
+            if (header.equals("delete")) {
+
+                boolean isDelete = pollService.deletePoll(id);
+                obj.put("deleted", isDelete);
+
+            } else if (header.equals("activate")) {
+
+                boolean isActivated = pollService.updatePollStatus(id, "Active");
+                obj.put("isUpdated", isActivated);
+            }
+            if (header.equals("deactivate")) {
+
+                boolean isDeactivated = pollService.updatePollStatus(id, "Inactive");
+                obj.put("isUpdated", isDeactivated);
+            }
+
             response.getWriter().write(obj.toString());
-            
+
         } catch (SQLException | ClassNotFoundException ex) {
             Logger.getLogger(ProcessPollsController.class.getName()).log(Level.SEVERE, null, ex);
         }
-        
+
     }
 }
